move AutoEvm and AutoId page component in Domains/
